+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: teppei <teppei@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/06/30 19:13:52 by tkitagaw          #+#    #+#             */
-/*   Updated: 2021/10/09 23:23:42 by teppei           ###   ########.fr       */
+/*   Updated: 2021/12/25 17:19:28 by teppei           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -98,6 +98,6 @@ char			**ft_ptrdup(char **strs);
 void			swap_str(char **strs, int i, int j);
 void			ft_bubble_sort_ascii(char **strs);
 void			ft_perror(char *str, char *err);
-int				ft_abs(int a);
+long			ft_abs(long a);
 
 #endif
